Add chat, question, slide update realtime

+++ src/components/ClassCard/ClassCard.tsx
@@ -1,20 +1,24 @@
 import { useNavigate } from 'react-router-dom';
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import styles from './styles.module.css';
 import { IconButton, Menu, MenuItem, Avatar } from '@mui/material';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
 import ClassThunks from '@/redux/feature/class/thunk';
 import clsx from 'clsx';
+
 import ClassApi from '@/api/classApi';
+import { useAppSelector } from '@/redux';
+import UserApi from '@/api/userApi';
 
 type Props = {
   name: string;
-  creatorName: string;
+  creatorId: string;
   subjectName: string;
   uuid: string;
 };
 
-function ClassCard({ name, creatorName, subjectName, uuid }: Props) {
+function ClassCard({ name, creatorId, subjectName, uuid }: Props) {
+  const loggedUser = useAppSelector(state => state.user.data);
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
   const open = Boolean(anchorEl);
   const handleClick = (event: React.MouseEvent<HTMLElement>) => {
@@ -30,46 +34,61 @@ function ClassCard({ name, creatorName, subjectName, uuid }: Props) {
   };
 
   const handleDeleteClass = async (uuid: string) => {
+    console.log("Trigger delete class.");
     handleClose();
+
+    if (creatorId != loggedUser.uuid) {
+      alert("You don't own this class");
+      return;
+    }
+    
     const payload = {
         uuid: uuid
     }
     ClassApi.deleteClass(uuid);
     //const response = await dispatcher(ClassThunks.);
     //console.log(response);
-}
+  }
+
+  const [creator, setCreator] = useState<any>({
+    uuid: "",
+    fullName: "",
+    email: "",
+    age: "",
+    dob: "",
+  });
+  const fetchCreator = async () => {
+    const response = await UserApi.getUserById(creatorId);
+    console.log(response);
+    setCreator(response);
+  };
+  useEffect(() => {
+    fetchCreator();
+  }, []);
 
   return (
     <div className={styles.container} key={uuid}>
       <div>
         <div className={styles.image}></div>
-        <div className={styles.more}>
-          <IconButton
-            aria-label="more"
-            id="long-button"
-            aria-controls={open ? 'long-menu' : undefined}
-            aria-expanded={open ? 'true' : undefined}
-            aria-haspopup="true"
-            onClick={handleClick}
-          >
-            <MoreVertIcon />
-          </IconButton>
-          <Menu
-            id="long-menu"
-            MenuListProps={{
-              'aria-labelledby': 'long-button',
-            }}
-            anchorEl={anchorEl}
-            open={open}
-            onClose={handleClose}
-          >
-            <MenuItem onClick={() => handleDeleteClass(uuid)} disableRipple>
-              Delete
-            </MenuItem>
-          </Menu>
-        </div>
+        {(creatorId == loggedUser.uuid) && 
+          <div className={styles.more}>
+              <Menu
+                id="long-menu"
+                MenuListProps={{
+                  'aria-labelledby': 'long-button',
+                }}
+                anchorEl={anchorEl}
+                open={open}
+                onClose={handleClose}
+              >
+                <MenuItem onClick={() => handleDeleteClass(uuid)} disableRipple>
+                  Delete
+                </MenuItem>
+              </Menu>
+          </div>
+        }
         <div className={styles.name}>{name}</div>
-        <div className={clsx(styles.creatorName, styles.desc)}>{`Người tạo: ${creatorName}`}</div>
+        <div className={clsx(styles.creatorName, styles.desc)}>{`Người tạo: ${creator.fullName}`}</div>
         <div className={styles.desc}>{`Môn: ${subjectName}`}</div>
       </div>
